Build the physical hierarchy from the logical one
createRoot, both createCluster overloads, createFlatCluster, addModuleLeafInstsToCluster,
setClusterMetrics and treatEachMacroAsSingleCluster.

A cluster is not a module, and two rules change the tree's SHAPE rather than merely its
contents. A module whose metrics are empty gets no cluster and consumes no id -- and
"empty" counts instances only, so a module holding nothing but ignored cells is skipped
even though its area is not zero. A glue-logic cluster with no leaves is built, found
empty and discarded; one that survived would take an id, occupy a slot, contribute
nothing, and shift every later id.

Both names are load-bearing. A module cluster takes the module's HIERARCHICAL name,
because two modules sharing a leaf name in different branches would otherwise collide.
A glue cluster is named after its parent in parentheses, and the parentheses nest, which
is why a real hierarchy reads ((root)_glue_logic_0_1)_glue_logic_1_0.

Also pinned: glue takes a module's own instances only, since its children get their own
clusters and descending would double-count them; cluster metrics count leaves PLUS every
module held, not either; and a macro-only design types each cluster HardMacro, which
masks its standard-cell count downstream.

Mutation testing found a test that could not fail. One claiming to check the type mask
built a cluster whose metrics already held zero standard cells, so the masked accessor
read zero whether the mask applied or not. It now asserts what it can actually see, and
says why -- the mask is tested where it is observable, in the cluster tests.

The mutation table gained a rule it had already broken twice: no double quotes in a
pattern, because each row is one bash double-quoted string.

74 mutations, 0 holes. 126 tests, nine of which need a database.

+++ src/cluster.rs
@@ -10,8 +10,8 @@
 use std::collections::VecDeque;
 
 pub type ClusterId = i32;
-pub type InstId = u32;
-pub type ModuleId = u32;
+pub type InstId = usize;
+pub type ModuleId = usize;
 
 /// Upstream `enum ClusterType`.
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
